+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - fix `MmdAmmoPhysics` rigidbody initialization
 
+- fix uv morph load error
+
 ## 0.42.1
 
 - fix physics body transform not applied to bone when models bone are not sorted by index